+++ history_request.go
@@ -91,16 +91,6 @@ func (b *historyBuilder) IncludeMeta(withMeta bool) *historyBuilder {
 	return b
 }
 
-// func (b *historyBuilder) IncludeUUID(includeUUID bool) *historyBuilder {
-// 	b.opts.IncludeUUID = includeUUID
-// 	return b
-// }
-
-// func (b *historyBuilder) IncludeMessageType(includeMessageType bool) *historyBuilder {
-// 	b.opts.IncludeMessageType = includeMessageType
-// 	return b
-// }
-
 // QueryParam accepts a map, the keys and values of the map are passed as the query string parameters of the URL called by the API.
 func (b *historyBuilder) QueryParam(queryParam map[string]string) *historyBuilder {
 	b.opts.QueryParam = queryParam
@@ -129,12 +119,10 @@ type historyOpts struct {
 
 	Channel string
 
-	Start              int64
-	End                int64
-	QueryParam         map[string]string
-	WithMeta           bool
-	IncludeUUID        bool
-	IncludeMessageType bool
+	Start      int64
+	End        int64
+	QueryParam map[string]string
+	WithMeta   bool
 
 	// default: 100
 	Count int
@@ -204,8 +192,6 @@ func (o *historyOpts) buildQuery() (*url.Values, error) {
 	q.Set("reverse", strconv.FormatBool(o.Reverse))
 	q.Set("include_token", strconv.FormatBool(o.IncludeTimetoken))
 	q.Set("include_meta", strconv.FormatBool(o.WithMeta))
-	// q.Set("include_message_type", strconv.FormatBool(o.IncludeMessageType))
-	// q.Set("include_uuid", strconv.FormatBool(o.IncludeUUID))
 
 	SetQueryParam(q, o.QueryParam)